+++ src/models/Map.ts
@@ -1,0 +1,7 @@
+import mapboxgl from 'mapbox-gl'
+
+export interface MapConfig {
+  sourceMapId: string
+  center: mapboxgl.LngLatLike
+  zoom: number
+}